Split testing into separate files for weather and sensor entities (as per review suggestion)
